transport stronger

+++ src/main/java/com/beyond/jgit/storage/SardineStorage.java
@@ -1,9 +1,13 @@
 package com.beyond.jgit.storage;
 
+import com.beyond.jgit.util.JsonUtils;
+import com.beyond.jgit.util.ObjectUtils;
 import com.beyond.jgit.util.PathUtils;
 import com.thegrizzlylabs.sardineandroid.Sardine;
 import com.thegrizzlylabs.sardineandroid.impl.OkHttpSardine;
 import com.thegrizzlylabs.sardineandroid.impl.SardineException;
+import lombok.Data;
+import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.StringUtils;
@@ -13,7 +17,9 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
 import java.util.*;
+import java.util.stream.Collectors;
 
+@Slf4j
 public class SardineStorage extends AbstractStorage {
 
     private final String basePath;
@@ -22,15 +28,18 @@ public class SardineStorage extends AbstractStorage {
 
     private final ExistDirCacheManager existDirs;
 
+    private final String sessionDir;
+
     public SardineStorage(String basePath, String username, String password) {
-        this(basePath, username, password, null);
+        this(basePath, username, password, null, null);
     }
 
-    public SardineStorage(String basePath, String username, String password, String existDirCachePath) {
+    public SardineStorage(String basePath, String username, String password, String existDirCachePath, String sessionDir) {
         this.basePath = basePath;
         sardine = new LoggedSardine(new OkHttpSardine());
         sardine.setCredentials(username, password);
         existDirs = new ExistDirCacheManager(existDirCachePath);
+        this.sessionDir = sessionDir;
     }
 
     @Override
@@ -47,6 +56,21 @@ public class SardineStorage extends AbstractStorage {
         String absPath = getAbsPath(targetPath);
         forceMkdirParent(absPath);
         sardine.put(absPath, file, null);
+    }
+
+    @Override
+    public void uploadBatch(List<TransportMapping> mappings) throws IOException {
+        Session session = new Session(sessionDir, Session.generateSessionId(Session.TransportType.upload,mappings));
+        session.start(mappings);
+        for (TransportMapping mapping : mappings) {
+            if (session.isDone(mapping)){
+                log.info("uploaded, no upload again. if something is wrong, delete session:{}", PathUtils.concat(sessionDir,session.getSessionId()));
+                continue;
+            }
+            upload(new File(mapping.getLocalPath()), mapping.getRemotePath());
+            session.done(mapping);
+        }
+        session.complete();
     }
 
     private void forceMkdirParent(String absPath) throws IOException {
@@ -140,6 +164,9 @@ public class SardineStorage extends AbstractStorage {
         }
 
         public void add(String relativePath) throws IOException {
+            if (existDirs.contains(relativePath)) {
+                return;
+            }
             if (StringUtils.isNotBlank(cachePath)) {
                 FileUtils.writeLines(new File(cachePath), Collections.singletonList(relativePath), true);
             }
@@ -151,10 +178,68 @@ public class SardineStorage extends AbstractStorage {
         }
 
         public void delete(String path) throws IOException {
-            if (contains(path)){
+            if (contains(path)) {
                 existDirs.remove(path);
                 FileUtils.writeLines(new File(cachePath), existDirs);
             }
+        }
+    }
+
+    @Data
+    private static class Session {
+
+        private String sessionDir;
+        private String sessionId;
+
+        public Session(String sessionDir, String sessionId) {
+            this.sessionDir = sessionDir;
+            this.sessionId = sessionId;
+        }
+
+        public static String generateSessionId(TransportType transportType, List<TransportMapping> transportMappings) {
+            List<TransportMapping> sorted = transportMappings.stream().sorted(Comparator.comparing(x -> x.getLocalPath() + " " + x.getRemotePath())).collect(Collectors.toList());
+            return ObjectUtils.sha1hash((transportType.toString()+" "+JsonUtils.writeValueAsString(sorted)).getBytes());
+        }
+
+        @SuppressWarnings("ResultOfMethodCallIgnored")
+        public void start(List<TransportMapping> transportMappings) throws IOException {
+            File thisSessionDir = new File(PathUtils.concat(sessionDir, sessionId));
+            if (!thisSessionDir.exists()){
+                thisSessionDir.mkdirs();
+            }
+            for (TransportMapping transportMapping : transportMappings) {
+                File file = new File(PathUtils.concat(sessionDir, sessionId, transportMapping.toSha1hash()));
+                if (!file.exists()){
+                    file.createNewFile();
+                }
+            }
+        }
+
+        public void done(TransportMapping mapping) throws IOException {
+            File file = new File(PathUtils.concat(sessionDir, sessionId, mapping.toSha1hash()));
+            if (file.exists()){
+                FileUtils.forceDelete(file);
+            }
+        }
+
+        public boolean isDone(TransportMapping mapping) {
+            // 根据mapping的sha1hash对应路径的文件是否存在来判断是否已经传输完成
+            File thisSessionDir = new File(PathUtils.concat(sessionDir, sessionId));
+            if (thisSessionDir.exists()){
+                return !new File(PathUtils.concat(sessionDir, sessionId, mapping.toSha1hash())).exists();
+            }
+            return false;
+        }
+
+        public void complete() throws IOException {
+            File thisSessionDir = new File(PathUtils.concat(sessionDir, sessionId));
+            if (thisSessionDir.exists()){
+                FileUtils.forceDelete(thisSessionDir);
+            }
+        }
+
+        private enum TransportType{
+            upload,download
         }
     }
 }
